+++ clickhouse_driver/result.py
@@ -1,4 +1,5 @@
 from .progress import Progress
+from .protocol import ServerPacketTypes
 
 
 class QueryResult(object):
@@ -113,3 +114,12 @@ class IterQueryResult(object):
 
     # For Python 3.
     __next__ = next
+
+
+class QueryInfo(object):
+    def __init__(self):
+        self.profile_info = None
+
+    def store(self, packet):
+        if packet.type == ServerPacketTypes.PROFILE_INFO:
+            self.profile_info = packet.profile_info